weather-forcating-app-azure/requirements.txt
Flask==2.2.3
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
Flask-JWT-Extended==4.4.4
psycopg2-binary==2.9.5
requests==2.28.2
python-dotenv==1.0.0
gunicorn==20.1.0
redis==4.5.1
celery==5.2.7
prometheus-flask-exporter==0.22.3
pytest==7.3.1
pytest-flask==1.2.0
flake8==6.0.0
black==23.3.0
scikit-learn==1.2.2
pandas==1.5.3
matplotlib==3.7.1
pyyaml==6.0
kubernetes==26.1.0
azure-identity==1.12.0
azure-mgmt-containerservice==21.0.0
azure-mgmt-containerregistry==10.1.0